third_party/blink/web_tests/platform/win7/virtual/media-capture-image-ptz/external/wpt/mediacapture-image/MediaStreamTrack-getSettings-expected.txt

This is a testharness.js-based test.
PASS exercises MediaStreamTrack.getSettings() with PTZ permission denied
PASS MediaStreamTrack-getSettings
PASS exercises MediaStreamTrack.getSettings() with PTZ permission granted
FAIL MediaStreamTrack-getSettings 1 assert_equals: expected (number) 5 but got (undefined) undefined
Harness: the test ran to completion.
